requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitarray==2.9.2
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.1
cryptography==42.0.5
cytoolz==0.12.3
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
ecdsa==0.19.0
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.2.2
eth-utils==4.1.0
eth_abi==5.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
fastapi-pagination==0.12.23
fastapi_permissions==0.2.7
fastjsonschema==2.19.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hexbytes==0.3.1
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
lru-dict==1.2.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.0
pandocfilters==1.5.1
parsimonious==0.10.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.1
priority==2.0.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pyunormalize==15.1.0
pyzmq==26.0.3
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
taskgroup==0.0.0a4
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.4
